requirements.txt
Django==1.8.4
Pillow==2.9.0
beautifulsoup4==4.4.0
requests==2.7
djangorestframework==3.1.2
django-bootstrap3==6.2.2
django-filter==0.11
psycopg2==2.6.1
gunicorn==19.3.0
gevent==1.0.2